requirements.txt
espn-api
openai>=1.30.0
python-dotenv
flask
gunicorn
